backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
pydantic
SQLAlchemy
psycopg2-binary
python-dotenv
